refactor(a1): move require_auth + AUTH_DISABLED to core/auth.py (BR-CODE-1 slice 5)
Moves the security-critical auth gate out of app.py into core/auth.py
(verbatim): require_auth (Supabase-JWT + email-domain check), the
AUTH_DISABLED kill-switch, and _is_production. app.py re-exports all three
(app.require_auth IS core.auth.require_auth), so every app.<NAME> read +
Depends(require_auth) usage is unchanged.

Test impact is minimal: require_auth reads AUTH_DISABLED as a core.auth module
global, so test_security_hardening's one AUTH_DISABLED monkeypatch is repointed
to core.auth; the Supabase-lookup stub patches the shared requests module
(reaches core.auth unchanged). Security suite 6/6 (all 4 auth-gate cases:
accept confirmed, reject 401 audience / 403 unconfirmed / 403 foreign domain)
+ 122 route net = 128 green; entrypoint imports clean.
